Better reading of buffers, data holder
There were some problems with my fancy detection of event buffers,
so I switched to something that should be more robust. A new class
`Data` is now available to handle extracting and holding the
information from a single buffer. This will be used in conjunction
with `Experiment` as the organizational structure.