Using DetectorArray
Since the Si detector for the experiment is really a 16-strip silicon
detector, instead of having that be considered a single detector with
the complications arising from that, I've broken a single detector out
and am having the data hold an array of detectors.

This allows energy calibrations and histogramming to be done a little
easier for the detectors, which is nice.