Plotting
While looking at the calibration spectra, I noticed that a few of the
channels showed hundreds of counts within the rightmost bin. I don't
know if there were overflows that weren't marked as such in the stream,
or overflows that were just dumped into the last bin. So, to ensure that
the lowest and highest bins actually show nothing, since nothing should
be there, just set them to zero.