Buffer determination changed
Can use `buffer_type='evt|chn'` to set the buffer type instead of
supplying a buffer file. This is only necessary when reading from an
HDF5 file, since buffer type also automatically sets the detector type.

Instead of breaking, if the buffer type isn't determined (say, by not
passing in a buffer_type), it displays a warning so that the user can
set filenames, create the detector layout they want, pass in a buffer,
and other manual intervention.