Further changes
I need to rethink what I'm doing. `pyne.Experiment` has not been useful
as I've worked on some of the analysis, and it's almost better to have
`pyne.Data` as the highest level organization. The routines and classes
with `sap` would handle what I had delegated and not used in
`pyne.Experiment`.

But, actually doing the analysis comes first, so work on getting that
done then clean this up.