Fixed binning, default plotting 'steps', note for analyze
After looking at the energy calibration data, the tails we see with
high count rates are based on the poor detector response. We need to
include those counts as events, so we won't need our windowing.

What we will need is identification of possible proton peaks (seen
visually in some spectra) and, if so, count up those lower counts and
call them protons and not alphas.